requirements.txt
Django==4.2.7
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
websockets==11.0.3
pytest==7.4.3
pytest-django==4.5.2
pytest-asyncio==0.21.1
redis==5.0.1
aiohttp==3.8.6